chore: separate FileCard
