+++ src/FileCard/FileCard.tsx
@@ -1,9 +1,7 @@
 import { FC } from 'react'
-import styled, { css, keyframes } from 'styled-components'
 import { Icon, IconType } from '../Icon'
-import { Button } from '../Button'
 import { Spacer } from '../Layout/Spacer'
-import getShimmerStyles from '../helpers/getShimmerStyles'
+import * as Styled from './FileCard.styled'
 
 const getFileSizeString = (size: number) => {
   // get file size in kb, mb, gb
@@ -25,129 +23,6 @@ const getFileSizeString = (size: number) => {
 
   return `${size} B`
 }
-
-const messageAnimation = keyframes`
-  from {
-    scale: 0.5;
-    translate: 0 -50%;
-    opacity: 0;
-  }
-  to {
-    scale: 1;
-    opacity: 1;
-    translate: 0 -100%;
-  }
-`
-
-interface StyledFileCardProps {
-  $isFetching: boolean
-  $message?: string
-}
-
-const StyledFileCard = styled.div<StyledFileCardProps>`
-  background-color: var(--color-grey-01);
-  display: flex;
-  border-radius: 4px;
-  padding: 2px;
-  align-items: center;
-  cursor: pointer;
-  gap: 4px;
-  position: relative;
-
-  :hover {
-    background-color: var(--color-grey-02);
-  }
-
-  .title {
-    margin-left: 4px;
-  }
-
-  .length {
-    opacity: 0.5;
-    white-space: nowrap;
-  }
-
-  .size {
-    opacity: 0.5;
-    margin: 0 4px;
-    white-space: nowrap;
-  }
-
-  .icon {
-    font-size: 24px;
-  }
-
-  img {
-    width: 24px;
-    height: 24px;
-    object-fit: cover;
-    border-radius: 4px;
-  }
-
-  ${({ $isFetching }) =>
-    $isFetching &&
-    css`
-      user-select: none;
-      pointer-events: none;
-
-      & > * {
-        opacity: 0.5;
-      }
-
-      ${getShimmerStyles(undefined, undefined, {
-        // random number between 0.2 and 1.5
-        delay: Math.random() * (1.5 - 0.2) + 0.2,
-        speed: 1,
-      })}
-    `}
-
-  ${({ $message }) =>
-    $message &&
-    css`
-      &:hover {
-        /* little triangle */
-        &::before {
-          content: '';
-          position: absolute;
-          translate: 0 -100%;
-          top: -8px;
-          right: 12px;
-          border: 4px solid transparent;
-          border-top-color: var(--color-hl-01);
-          border-right-color: var(--color-hl-01);
-          rotate: 135deg;
-          animation: ${messageAnimation} 100ms ease-in-out;
-          transform-origin: bottom;
-        }
-
-        &::after {
-          content: '${$message}';
-          position: absolute;
-          top: -4px;
-
-          padding: 4px;
-          border-radius: 4px;
-          font-size: 12px;
-          right: -8px;
-          translate: 0 -100%;
-          animation: ${messageAnimation} 100ms ease-in-out;
-          transform-origin: right;
-          background-color: var(--color-hl-01);
-          color: black;
-        }
-      }
-    `}
-`
-
-const StyledButton = styled(Button)`
-  background-color: unset;
-  padding: 2px;
-  min-height: unset;
-
-  .icon {
-    color: inherit;
-  }
-`
 
 const fileIcons: {
   [key: string]: IconType
@@ -206,13 +81,13 @@ export const FileCard: FC<FileCardProps> = ({
     'insert_drive_file'
 
   return (
-    <StyledFileCard $isFetching={isFetching} {...props} $message={message}>
+    <Styled.FileCard $isFetching={isFetching} {...props} $message={message}>
       {preview ? <img src={preview} /> : <Icon icon={typeIcon} />}
       <span className="title">{title}</span>
       {length > 1 && <span className="length">{`(${length} files)`}</span>}
       <Spacer />
       <span className="size">{getFileSizeString(size)}</span>
-      <StyledButton
+      <Styled.Icon
         icon="vertical_split"
         onClick={onSplit}
         style={{
@@ -224,7 +99,7 @@ export const FileCard: FC<FileCardProps> = ({
         disabled={splitDisabled || isFetching || disabled}
       />
       {!readOnly && (
-        <StyledButton icon="close" onClick={onRemove} disabled={isFetching || disabled} />
+        <Styled.Icon icon="close" onClick={onRemove} disabled={isFetching || disabled} />
       )}
       {readOnly && (
         <Icon
@@ -234,6 +109,6 @@ export const FileCard: FC<FileCardProps> = ({
           }}
         />
       )}
-    </StyledFileCard>
+    </Styled.FileCard>
   )
 }
